moves the FormInputProps definition to reactive-form types
